pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "newsapi-python>=0.2.7",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "pycoingecko>=3.2.0",
    "streamlit>=1.42.0",
    "textblob>=0.19.0",
]
